+++ src/body/text.ts
@@ -16,4 +16,6 @@ export class TextBody implements IBody {
   public write = (response: ServerResponse) => {
     response.write(this.body);
   };
+
+  public toString = () => this.body;
 }